requirements.txt
streamlit
numpy
opencv-python
pandas
matplotlib
scikit-learn
tqdm
tensorflow-macos==2.16.2
tensorflow-metal==1.2.0
urllib3<2
